I updated the video.py to run on yolov5  instead of yolov4.
I really liked this PyTorch Hrnet implementation because it was relatively easier to integrate Yolo v5 through torch hub so I forked this repo and put a few lines of code to extract bboxes and send them to the posenet model! 

I also changed a few lines in the pose_util files to accommodate my YOLO v5 output.